refactor: trim and rewrite analytics plugins
